update: change license to MIT
GPLv3 is nice, but MIT just has extra freedom that I feel people (including companies) deserve. I don't care if someone uses this project in a commercial capacity, because this project will (hopefully) be a complete project that does everything it needs to do. Additionally, the character JSON should definitely be open for anyone, due to it's massive potential for usage in other projects.